DDA2-48: Request types dinamicos, con entidades, controller y services. Creación de esquemas para tablas en flyway.

+++ src/main/java/com/reclamos/backend/repository/RequestTypeRepository.java
@@ -3,5 +3,5 @@ package com.reclamos.backend.repository;
 import com.reclamos.backend.entity.RequestType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestTypeRepository extends JpaRepository<RequestType, Integer> {
+public interface RequestTypeRepository extends JpaRepository<RequestType, Long> {
 }
